refactor(ui-react-utils): fix provider export name
Since the export name was not the same as the filename babel-plugin-transform-imports could not
rewrite it to a local path, thus it could not be found for commonJS builds used by e.g. Jest
