+++ packages/ui-react-utils/src/DeterministicIdContext/generateInstanceCounterMap.ts
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-import type { DeterministicIdProviderValue } from './DeterministicIdProvider'
+import type { DeterministicIdProviderValue } from './DeterministicIdContextProvider'
 function generateInstanceCounterMap(): DeterministicIdProviderValue {
   return new Map<string, number>()
 }
